Add functions to map IR back to source code
To map IR back to source code we are using debug information provided by
the front end, i.e. IR file needs to be compiled with -g flag.
Depending on the IR instruction a subset of the following source
code information can be provided:
- name of the variable
- line and column
- path to source file (.cpp/.c)
- name of the enclosing function

This information allows the analysis writer to generate more expressive
reports at the end of an analysis run.