Implement report function for IFDSConstAnalysis
IFDSConstAnalysis will now report all immutable/const memory locations
using the proper report function. Source code information is also
included into the result report.

Additionally, a function that returns all methods of a IR Module was
added to the LLVMBasedICFG, and a function which gathers all stack and
heap memory locations was added to the ProjectIRDB.