Remove gtest from link list.  Add custom library path during install.
phasar was linked against gtest. On a stock 18.04 it was linking
dynamically, however this doesn't appear to be necessary and creates a
runtime dependency requiring libgtest to be installed.  I left these
changes commented hoping for clarification.  In particular, it was
noticed that in the sample VM libgtest.so doesn't seem to exist
anywhere.

Also, because of changes I made to install phasar to a custom directory
I need to add the library path to the library cache in the bootstrap.sh
script.