Refactor Points-To graph merging logic.
Changes the merge logic for points-to graph to avoid re-creating the
ValueVertexMap from scratch each time.  When merging we can ask boost to
provide a mapping so we can simply append to the existing ValueVertexMap
instead.  For large applications this makes a measurable runtime
improvement since the number of merge calls is large.