Optimize calls to get the Users for a a points-to vertex.
The call to Value->users() is quite expensive.  Because the result
should not change between executions, we can cache the results for each
Vertex of the points-to graph.  This has a significant performance
improvement.

This change also modifies the n^2 loop to be n^2/2 as described in the
comments -- the inner loop was previously starting at the beginning of
the map instead of at the next item in the map, and we can avoid looking
up in the map by using the appropriate iterators.  These changes did not
significantly impact performance, however the resulting code is more
concise.