ICFG Optimizations -- graph merging and others.
The largest change performance wise is the change to the graph type to
use a vector for edge storage.

The merge function now mirrors the changes to the Points-To class to
avoid re-building the FunctionVertexMap from scratch.

Changes in constructionWalker avoid fetching the same data multiple
times back-to-back.

printAsDot was modified to allow the user to skip the edge labels,
because those are quite expensive.  Moved code out of the header that is
only used internally to this function body.


ProjectIRDB.h does not use the LLVMPointsToGraph.h, so it was removed
which necessitated several other header changes.