Fix some bugs in LLVMTypeHierarchy.cpp
in methods like removeStructOrClassPrefix, ltrim is used to remove a prefix
from a name, which may be wrong. According to the comments of ltrim, ltrim
removes all chars belonging to the prefix from the target string, instead
of removing the prefix directly. The right way is simply removing the prefix
using drop_front