Fix a bug in LLVMTypeHierarchy::getVirtualFunctions
An entry in the virtual function table may be a GlobalAlias,
in which case, the function can not be retrieved from the module
using the name of the GlobalAlias.
This patch fix the issue by getting the function using the name of
the its aliasee if the entry is a GlobalAlias.