Fixes linking issue, where PTS requires getNumGlobals
We need to move getNumGlobals into the ProjectIRDB header because the
current implementation of the PointsToSet constructor uses it, however,
phasar_db(ProjectIRDB) links against phasar_pointer, so we cannot link
PointsToSet against phasar_db. By defining the function in the header we
cirmumvent this issue, however, we should be careful to not introduce
cyclic dependencies in the future.